Work on zenodo-service DB

+++ zenodo-service/src/main/kotlin/dk/sdu/cloud/zenodo/api/ZenodoCommandStreams.kt
@@ -5,7 +5,7 @@ import dk.sdu.cloud.service.KafkaDescriptions
 data class ZenodoPublishCommand(
     val jwt: String,
     val uuid: String,
-    val publicationId: Int,
+    val publicationId: Long,
     val request: ZenodoPublishRequest
 )
